requirements.txt
appnope==0.1.2
async-generator==1.10
attrs==21.2.0
bleach==3.3.1
cachetools==4.2.2
certifi==2021.5.30
chardet==3.0.4
decorator==5.0.9
defusedxml==0.7.1
entrypoints==0.3
google-api-core==1.31.1
google-api-python-client==2.15.0
google-auth==1.34.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.53.0
httplib2==0.19.1
idna==2.8
ipykernel==4.6.1
ipython==5.5.0
ipython-genutils==0.2.0
Jinja2==3.0.1
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
MarkupSafe==2.0.1
mistune==0.8.4
nbclient==0.5.3
nbconvert==6.1.0
nbformat==5.1.3
nest-asyncio==1.5.1
numpy==1.21.1
oauth2client==4.1.3
packaging==21.0
pandocfilters==1.4.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.3.1
portpicker==1.2.0
prompt-toolkit==1.0.18
protobuf==3.17.3
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyDrive==1.3.1
Pygments==2.9.0
pyparsing==2.4.7
pyrsistent==0.18.0
python-dateutil==2.8.2
pytz==2021.1
PyYAML==5.4.1
pyzmq==22.1.0
requests==2.21.0
rsa==4.7.2
simplegeneric==0.8.1
six==1.16.0
terminado==0.10.1
testpath==0.5.0
tornado==4.5.3
traitlets==5.0.5
uritemplate==3.0.1
urllib3==1.24.3
wcwidth==0.2.5
webencodings==0.5.1
